+++ activejob/lib/active_job/queue_adapters/sidekiq_adapter.rb
@@ -17,20 +17,24 @@ module ActiveJob
     class SidekiqAdapter
       def enqueue(job) #:nodoc:
         #Sidekiq::Client does not support symbols as keys
-        Sidekiq::Client.push \
+        sidekiq_job_id = Sidekiq::Client.push \
           'class'   => JobWrapper,
           'wrapped' => job.class.to_s,
           'queue'   => job.queue_name,
           'args'    => [ job.serialize ]
+        job.provider_job_id = sidekiq_job_id
+        sidekiq_job_id
       end
 
       def enqueue_at(job, timestamp) #:nodoc:
-        Sidekiq::Client.push \
+        sidekiq_job_id = Sidekiq::Client.push \
           'class'   => JobWrapper,
           'wrapped' => job.class.to_s,
           'queue'   => job.queue_name,
           'args'    => [ job.serialize ],
           'at'      => timestamp
+        job.provider_job_id = sidekiq_job_id
+        sidekiq_job_id
       end
 
       class JobWrapper #:nodoc: