+++ activejob/test/integration/queuing_test.rb
@@ -61,4 +61,14 @@ class QueuingTest < ActiveSupport::TestCase
     test_job = TestJob.perform_later @id
     assert_kind_of Fixnum, test_job.provider_job_id
   end
+
+  test 'should supply a provider_job_id to Sidekiq' do
+    skip unless adapter_is?(:sidekiq)
+    test_job = TestJob.perform_later @id
+    refute test_job.provider_job_id.nil?, "Provider job id should be set by Sidekiq"
+
+    delayed_test_job = TestJob.set(wait: 1.minute).perform_later @id
+    refute delayed_test_job.provider_job_id.nil?,
+      "Provider job id should by set for delayed jobs by sidekiq"
+  end
 end